cowork/core_api/pyproject.toml
[project]
name = "cowork-core-api"
version = "0.1.0"
description = "Core API service for persistence and account management"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
]

[tool.ruff]
line-length = 120
